src/main/java/com/github/k24/retrofit2/converter/success/SuccessConverterFactory.java
/*
 * Copyright 2017 k24
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.github.k24.retrofit2.converter.success;

import okhttp3.RequestBody;
import okhttp3.ResponseBody;
import retrofit2.Converter;
import retrofit2.Retrofit;

import java.io.IOException;
import java.lang.annotation.Annotation;
import java.lang.reflect.Type;

/**
 * Created by k24 on 2017/04/11.
 */
public class SuccessConverterFactory extends Converter.Factory {
    public static SuccessConverterFactory create() {
        return new SuccessConverterFactory();
    }

    private SuccessConverterFactory() {
    }

    @Override
    public Converter<ResponseBody, ?> responseBodyConverter(Type type, Annotation[] annotations, Retrofit retrofit) {
        if (!Success.class.equals(type)) return null;
        return new Converter<ResponseBody, Success>() {
            @Override
            public Success convert(ResponseBody responseBody) throws IOException {
                return Success.SUCCESS;
            }
        };
    }

    @Override
    public Converter<?, RequestBody> requestBodyConverter(Type type, Annotation[] parameterAnnotations, Annotation[] methodAnnotations, Retrofit retrofit) {
        return null;
    }
}
